requirements.txt
dm_env==1.6
graphviz==0.20.1
gym==0.19.0
gymnasium==0.29.1
ipdb==0.13.9
matplotlib==3.7.1
metaworld==2.0.0
networkx==3.0
numpy==1.23.5
pandas==1.4.2
pgmpy==0.1.25
pydot==1.4.2
pygam==0.9.1
pygraphviz==1.13
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.13.0
setuptools==65.6.3
statsmodels==0.14.0
torch==1.11.0
tqdm==4.64.0
wandb==0.13.10
